requirements.txt
streamlit==1.54.0
streamlit-drawable-canvas==0.9.3
tensorflow==2.20.0
keras==3.13.2
numpy==1.26.4
pillow==10.4.0
